requirements.txt
django>=4.2
djangorestframework>=3.14
python-docx>=0.8.11
pypandoc>=1.11
